Langserve/requirements.txt
langchain>=0.0.148
langserve>=0.1.2
fastapi>=0.95.1
uvicorn>=0.22.0
python-dotenv>=1.0.0
openai>=0.27.0